Created CoinDataService

+++ CryptoTracker/CryptoTracker/Models/Coin.swift
@@ -12,7 +12,7 @@ struct Coin: Identifiable, Codable {
     let image: String
     let currentPrice: Double
     let marketCap, marketCapRank, fullyDilutedValuation: Double?
-    let totalVolume, high24H, low24H: Int?
+    let totalVolume, high24H, low24H: Double?
     let priceChange24H: Double?
     let priceChangePercentage24H: Double?
     let marketCapChange24H: Double?
